requirements.txt
Flask==2.0.3
python-gitlab==2.10.1
requests==2.26.0
requests-toolbelt==0.9.1
urllib3==1.22
